linker: Allow the shim lib to overwrite symbols in its target lib
Current DT_NEEDED mimick method did not allow overriding symbols in a
lib, but only adding new symbols. This is because the shim lib would
have been in a different namespace from the target lib, where the
target lib takes priority over the shim lib, as it was DT_NEEDED by
the target lib.

Note that you may _NOT_ override symbols in an executable still, as in
the case of LD_PRELOAD. The shim lib still acts as a DT_NEEDED of the
exectuable.

Change-Id: I3f300263a9ca2b90c2ba8ebf30d32c2df20801ae